LayoutTests/fast/css/print-color-adjust-expected.txt

economy: PASS
exact: PASS
shorthand-economy: PASS
shorthand-exact: PASS
prefixed-economy: PASS
prefixed-exact: PASS
